@@ -25,13 +25,13 @@ setup(
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
         'Programming Language :: Python :: 3.12',
         'Programming Language :: Python :: 3.13',
     ],
+    python_requires='>=3.9',
     install_requires=[
         'girder>=5.0.0a2',
         'girder-jobs>=5.0.0a2',
@@ -46,7 +46,6 @@ setup(
     include_package_data=True,
     keywords='girder-plugin, large_image',
     packages=find_packages(exclude=['test', 'test.*', 'test_girder', 'test_girder.*']),
-    python_requires='>=3.8',
     url='https://github.com/girder/large_image',
     zip_safe=False,
     entry_points={
