@@ -25,13 +25,13 @@ setup(
     classifiers=[
         'Development Status :: 5 - Production/Stable',
         'Programming Language :: Python :: 3',
-        'Programming Language :: Python :: 3.8',
         'Programming Language :: Python :: 3.9',
         'Programming Language :: Python :: 3.10',
         'Programming Language :: Python :: 3.11',
         'Programming Language :: Python :: 3.12',
         'Programming Language :: Python :: 3.13',
     ],
+    python_requires='>=3.9',
     install_requires=[
         'jsonschema>=2.5.1',
         f'girder-large-image{limit_version}',
@@ -40,9 +40,8 @@ setup(
     extras_require={
         'compute': [
             'openpyxl',
-            'pandas ; python_version < "3.9"',
-            'pandas>=2.2 ; python_version >= "3.9"',
-            'python-calamine ; python_version >= "3.9"',
+            'pandas>=2.2',
+            'python-calamine',
             'umap-learn',
         ],
         'tasks': [
@@ -52,7 +51,6 @@ setup(
     include_package_data=True,
     keywords='girder-plugin, large_image',
     packages=find_packages(exclude=['test', 'test.*', 'test_annotation', 'test_annotation.*']),
-    python_requires='>=3.8',
     url='https://github.com/girder/large_image',
     zip_safe=False,
     entry_points={
