@@ -415,7 +415,7 @@ class AnnotationResource(Resource):
         if 'empty' in elementdict:
             elementdict.pop('empty')
             for el in elements:
-                elementdict[el['id']] = el
+                elementdict[str(el['id'])] = el
         if '/' in elpath:
             return self._patchEntry(elementdict, elpath, op, value, fullpath)
         elid = elpath.split('/', 1)[0].lower()
@@ -1122,11 +1122,16 @@ class AnnotationResource(Resource):
 
     @access.public(scope=TokenScope.DATA_READ)
     @autoDescribeRoute(
-        Description('Get annotation counts for a list of items.')
+        Description(
+            'Get annotation counts for a list of items.  If using actual a '
+            'database other than DocumentDB, this also indicates if items are '
+            'referenced as annotations.')
         .param('items', 'A comma-separated list of item ids.')
         .errorResponse(),
     )
     def getItemListAnnotationCounts(self, items):
+        from girder_large_image.models.image_item import ImageItem
+
         user = self.getCurrentUser()
         results = {}
         for itemId in items.split(','):
@@ -1135,10 +1140,11 @@ class AnnotationResource(Resource):
                 {'_active': {'$ne': False}, 'itemId': item['_id']},
                 user=self.getCurrentUser(), level=AccessType.READ, limit=-1)
             results[itemId] = annotations.count()
-            if Annotationelement().findOne({'element.girderId': itemId}):
-                if 'referenced' not in results:
-                    results['referenced'] = {}
-                results['referenced'][itemId] = True
+            if not ImageItem().checkForDocumentDB():
+                if Annotationelement().findOne({'element.girderId': itemId}):
+                    if 'referenced' not in results:
+                        results['referenced'] = {}
+                    results['referenced'][itemId] = True
         return results
 
     @access.user(scope=TokenScope.DATA_WRITE)
